models/02_plant_locator_v2/requirements.txt
affine==2.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
cython-bbox==0.1.5
dataclasses-json==0.6.4
defusedxml==0.7.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
geographiclib==2.0
humanfriendly==10.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
lap==0.4.0
loguru==0.7.2
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onemetric==0.1.2
onnxruntime==1.17.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
PyExifTool==0.5.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.9
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
snuggs==1.4.7
supervision==0.18.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.10
urllib3==2.2.0
